requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.20
future==0.18.2
geojson==2.5.0
geopandas==0.9.0
idna==2.10
lxml==4.6.3
munch==2.5.0
numpy==1.20.3
overpass==0.7
paho-mqtt==1.5.1
pandas==1.2.4
pymavlink==2.4.15
pyproj==3.1.0
pyreadline==2.1
PySide2==5.15.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Shapely==1.7.1
shiboken2==5.15.2
six==1.16.0
urllib3==1.26.5
